images/jobs/pageview-fetcher/requirements.txt
dlt==1.9.0
dlt[bigquery]
google-cloud-bigquery
google-auth
